tests/wpt/web-platform-tests/css/css-transforms-2/animation/rotate-interpolation.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>rotate interpolation</title>
    <link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
    <link rel="help" href="https://drafts.csswg.org/css-transforms-2/#propdef-rotate">
    <meta name="assert" content="rotate supports animation.">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="resources/interpolation-testcommon.js"></script>
  </head>
  <body>
    <script>
      test_interpolation({
        property: 'rotate',
        from: '100deg',
        to: '180deg',
      }, [
        {at: -1, expect: '20deg'},
        {at: 0, expect: '100deg'},
        {at: 0.125, expect: '110deg'},
        {at: 0.875, expect: '170deg'},
        {at: 1, expect: '180deg'},
        {at: 2, expect: '260deg'}
      ]);

      test_interpolation({
        property: 'rotate',
        from: '45deg',
        to: '-1 1 0 60deg',
      }, [
        {at: -1, expect: '0.447214 -0.447214 0.774597 104.478deg'},
        {at: 0, expect: '45deg'},
        {at: 0.125, expect: '-0.136456 0.136456 0.981203 40.6037deg'},
        {at: 0.875, expect: '-0.70246 0.70246 0.114452 53.1994deg'},
        {at: 1, expect: '-1 1 0 60deg'},
        {at: 2, expect: '-0.637897 0.637897 -0.431479 124.975deg'}
      ]);

      test_interpolation({
        property: 'rotate',
        from: 'none',
        to: '7 -8 9 400grad',
      }, [
        {at: -1, expect: '7 -8 9 -400grad'},
        {at: 0, expect: 'none'},
        {at: 0.125, expect: '7 -8 9 50grad'},
        {at: 0.875, expect: '7 -8 9 350grad'},
        {at: 1, expect: '7 -8 9 400grad'},
        {at: 2, expect: '7 -8 9 800grad'}
      ]);

      test_interpolation({
        property: 'rotate',
        from: 'none',
        to: 'none',
      }, [
        {at: -1, expect: 'none'},
        {at: 0, expect: 'none'},
        {at: 0.125, expect: 'none'},
        {at: 0.875, expect: 'none'},
        {at: 1, expect: 'none'},
        {at: 2, expect: 'none'}
      ]);
    </script>
  </body>
</html>
